Add new servlets for lights and scenes and update the lightcontrollerprovider to take a long instead of an int

+++ bcc-light-controller-core/src/main/java/com/s2d/bcc/light/controller/core/spi/LightControllerProvider.java
@@ -15,7 +15,7 @@ public abstract class LightControllerProvider
    */
   public abstract LightController openLightController ();
   
-  public abstract LightController openLightController ( int id );
+  public abstract LightController openLightController ( long id );
   
   public abstract Collection < Integer > getLightControllerIds ();
